app/requirements.txt
pandas==2.2.3
aiogram==3.14.0
python-dotenv==1.0.1
uvicorn==0.34.0
pymupdf4llm==0.0.17
openpyxl==3.1.5
# opencv-python==4.11.0.86
# scikit-image==0.25.1
# scikit-learn==1.6.1
rembg==2.0.62
# onnxruntime==1.19.2
# torch==2.6.0
# torchvision==0.21.0
cython==3.0.12
pycocotools==2.0.8
fastapi==0.115.11
openai==1.66.3
oauth2client== 4.1.3
gspread==6.2.0
reportlab==4.3.1
pycocotools==2.0.8
lunarcalendar==0.0.9